requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Stock data
yfinance>=0.2.36

# Embeddings (local, free)
sentence-transformers>=2.2.2
torch>=2.0.0

# Vector storage
chromadb>=0.4.22

# News parsing
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# CLI
typer>=0.9.0
rich>=13.0.0
